requirements.txt
altair==5.2.0
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.1.0
blinker==1.7.0
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
folium==0.15.1
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
htmlmin==0.1.12
idna==3.6
ImageHash==4.3.1
importlib-metadata==6.11.0
ipyleaflet==0.18.1
ipython==8.18.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
multimethod==1.10
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pandas-profiling==3.2.0
parso==0.8.3
pexpect==4.9.0
phik==0.12.3
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
smmap==5.0.1
st-pages
stack-data==0.6.3
streamlit==1.29.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
traittypes==0.2.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
visions==0.7.4
wcwidth==0.2.12
widgetsnbextension==4.0.9
xyzservices==2023.10.1